+++ docs/pydl/changes.rst
@@ -7,12 +7,16 @@ PyDL Changelog
 
 * Changes in how Mangle-polygon containing FITS files are handled, related to
   `Issue #11`_.
+* Fixed memory leak in :func:`~pydl.pydlspec2d.spec2d.combine1fiber`,
+  see `Issue #12`_.
 * Added :func:`~pydl.pydlutils.mangle.is_in_window`.
 * Allow polygon area functions to deal with negative caps and ``use_caps``.
 * Update ``docs/conf.py`` for Python 3.5 compatibility (`PR #13`_).
 
 .. _`PR #13`: https://github.com/weaverba137/pydl/pull/13
 .. _`Issue #11`: https://github.com/weaverba137/pydl/issues/11
+.. _`Issue #12`: https://github.com/weaverba137/pydl/issues/12
+
 
 0.5.1 (2016-06-22)
 ------------------